user list is shown dynamically
