+++ frontend/src/leave.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { withRouter } from 'react-router-dom';
 import axios from 'axios';
+import socket from './index'
 
 const Leave = (props) => {
 
@@ -10,6 +11,7 @@ const Leave = (props) => {
         axios
           .post('/api/user/leave', props.user)
           .then(function(response) {
+            socket.emit('retrieveUsers');
             console.log(response.data);
             props.history.push('/')
           })
