+++ socket.js
@@ -1,4 +1,4 @@
-module.exports = (app, messages, io) => {
+module.exports = (app, messages, users, io) => {
     io.on('connection', function(client) {
         console.log('Client connected...');
         // on receipt of new message, saves the message
@@ -12,6 +12,15 @@ module.exports = (app, messages, io) => {
         client.on('retrieveMessages', function() {
           client.emit('listOfMessages', messages.messageList);
         });
+
+        client.on('retrieveUsers', function () {
+            const allFakeNames = users.getAllFakeNames();
+            const allRealNames = users.getAllRealNames();
+          io.emit('listOfUsers', {
+            allFakeNames,
+            allRealNames
+          });
+        })
 
       });
 }
